requirements.txt
asgiref==3.3.4
Django==3.2
django-appconf==1.0.4
django-cors-headers==3.7.0
django-imagekit==4.0.2
django-redis==4.12.1
django-simpleui==2021.4.1
django-taggit==1.3.0
djangorestframework==3.12.4
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
mysqlclient==2.0.3
pilkit==2.0
Pillow==8.2.0
Pygments==2.8.1
pytz==2021.1
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
